2022: day21: part 2
Do I feel a bit silly for doing a full single variable algebra system
which can do multiplication and division of terms of arbitrary order?
Yes, a little bit.

Am I proud of doing it?
Yes, a lot.

I probably should have guessed that, if there was just **one** solution
for this problem, then it was just a simple linear / first degree
polynomial. And in that case I could have simplified everything
dramatically. But as is, this was tons of fun.